@koind 2022-08-08
**До оптимизации:** 
```
goos: darwin
goarch: amd64
pkg: github.com/koind/avito_otus_hw/hw10_program_optimization
cpu: Intel(R) Core(TM) i5-5257U CPU @ 2.70GHz
BenchmarkGetDomainStat-4   	       4	 274351485 ns/op	132845794 B/op	 1700092 allocs/op
PASS
ok  	github.com/koind/avito_otus_hw/hw10_program_optimization	3.715s
```


**После оптимизации:**
```
goos: darwin
goarch: amd64
pkg: github.com/koind/avito_otus_hw/hw10_program_optimization
cpu: Intel(R) Core(TM) i5-5257U CPU @ 2.70GHz
BenchmarkGetDomainStat-4   	  166515	      6735 ns/op	    5928 B/op	      41 allocs/op
PASS
ok  	github.com/koind/avito_otus_hw/hw10_program_optimization	1.595s
```


**benchstat:**
```
name                      old time/op                 new time/op       delta
GetDomainStat-4  274351.48µs ± 0%   6.74µs ± 0%   ~   (p=1.000 n=1+1)

name                      old alloc/op                new alloc/op       delta
GetDomainStat-4  132845.79kB ± 0%   5.93kB ± 0%  ~   (p=1.000 n=1+1)

name                      old allocs/op              new allocs/op     delta
GetDomainStat-4  1700092.0 ± 0%       41.0 ± 0%       ~   (p=1.000 n=1+1)
```


**benchcmp:**
```
benchmark                    old ns/op     new ns/op     delta
BenchmarkGetDomainStat-4     274351485     6735          -100.00%

benchmark                    old allocs     new allocs     delta
BenchmarkGetDomainStat-4     1700092        41             -100.00%

benchmark                    old bytes     new bytes     delta
BenchmarkGetDomainStat-4     132845794     5928          -100.00%
```